perf(orm): 重新调整数据库主键映射数据类型为int32

+++ server/internal/http/admin/controller/system/menu.go
@@ -66,7 +66,7 @@ func (c *Menu) Edit(w http.ResponseWriter, r *http.Request) {
 // Delete 删除分类
 func (c *Menu) Delete(w http.ResponseWriter, r *http.Request) {
 	req := struct {
-		Ids []int64 `json:"ids"`
+		Ids []int32 `json:"ids"`
 	}{}
 	if err := c.JsonReqUnmarshal(r, &req); err != nil {
 		_ = c.Fail(w, 101, "请求参数异常", err.Error())
