+++ server/internal/http/admin/controller/system/role.go
@@ -64,7 +64,7 @@ func (c *Role) Edit(w http.ResponseWriter, r *http.Request) {
 
 func (c *Role) Delete(w http.ResponseWriter, r *http.Request) {
 	req := struct {
-		Ids []int64 `json:"ids"`
+		Ids []int32 `json:"ids"`
 	}{}
 	if err := c.JsonReqUnmarshal(r, &req); err != nil {
 		_ = c.Fail(w, 101, "请求参数异常", err.Error())
