+++ server/internal/http/admin/controller/system/user.go
@@ -78,7 +78,7 @@ func (c *User) Edit(w http.ResponseWriter, r *http.Request) {
 
 func (c *User) Delete(w http.ResponseWriter, r *http.Request) {
 	req := struct {
-		Ids []int64 `json:"ids"`
+		Ids []int32 `json:"ids"`
 	}{}
 	if err := c.JsonReqUnmarshal(r, &req); err != nil {
 		_ = c.Fail(w, 101, "请求参数异常", err.Error())
