+++ server/internal/http/admin/dto/menu.go
@@ -5,28 +5,28 @@ import "app/internal/http/common/services"
 type MenuListReq struct {
 	Page     int    `form:"page" validate:"min:1" label:"分页"`
 	PageSize int    `form:"page_size" validate:"min:1|max:1000"`
-	AppId    int64  `form:"app_id" validate:"gt:0"`
+	AppId    int32  `form:"app_id" validate:"gt:0"`
 	Keyword  string `form:"keyword"`
 }
 
 type MenuCreateReq struct {
-	Pid       int64             `json:"pid"`
+	Pid       int32             `json:"pid"`
 	Name      string            `json:"name"`
 	Path      string            `json:"path"`
 	Component string            `json:"component"`
 	Meta      services.MenuMate `json:"meta"`
-	AppId     int64             `json:"app_id"`
+	AppId     int32             `json:"app_id"`
 }
 
 type MenuUpdateReq struct {
-	Id        int64             `json:"id"`
+	Id        int32             `json:"id"`
 	Name      string            `json:"name"`
 	Path      string            `json:"path"`
 	Component string            `json:"component"`
 	Sort      int32             `json:"sort"`
 	Meta      services.MenuMate `json:"meta"`
-	AppId     int64             `json:"appId"`
-	Pid       int64             `json:"pid"`
+	AppId     int32             `json:"appId"`
+	Pid       int32             `json:"pid"`
 	Status    int32             `json:"status"`
 	ApiList   []struct {
 		Path string `json:"path"`
